Update patch set 15
Patch Set 15:

> When rebasing please fix function name in commit description
 > (nvme_connect_slow -> nvme_connect)

Thanks for noticing!
I was planning to name it _slow to reflect that we do not hurry when using it.
So would that be ok with you if I change the function name in scripts instead?

Patch-set: 15
Reviewer: Vitaliy Mysak <1014935@d5d70762-12d0-45a1-890d-524b12d3f735>