Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

This code is implementing a flash translation layer used to emulate a block device on top of an open channel ssd. The library name should be something like "ftl" or "ocssd_ftl" so that it is clear what this library is doing.

Just naming it "ocssd" makes me think it's a library for interacting with OCSSDs in general, not specifically for using them as standard block devices.

Patch-set: 2
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1