Update patch set 2
Patch Set 2:

I also want to say that I really appreciate the effort to break these patches up a bit for us. That makes our lives much easier. However, when we say to break patches up, we mean into functional, self-contained units with a clear purpose. As a concrete example, instead of breaking the code up into two patches, one with the header files and one with the implementation files, you could have done the following:

1) First patch adds a lib/ocssd library that contains a single, mostly empty file. That patch includes all of the Makefile changes to build that library.
2) Second patch adds the initialization and shutdown paths for the library, including the public API functions and their implementations. These functions may have reduced functionality at first because nothing else is implemented, but they can allocate the base ocssd object and such.
3) Third patch adds the ocssd_dev structure and functions to create and destroy one.
4) Fourth patch adds the read path
5) Fifth patch adds the write path
6) Sixth patch adds garbage collection
etc.

Then when we review the patches, we can focus on reviewing one thing at a time. As it is now, I'm mostly pulling down the completed set of patches, reviewing it in the functional units I described above, and then searching for which patch adds the code and placing my comment there. This is mostly a note for the future - we'll proceed on the review here as quickly as we can.

Patch-set: 2
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>