Update patch set 2
Patch Set 2:

(5 comments)

As a precursor to this patch, is there any way you can write up a design document (in markdown and place it in the doc directory) that describes at a high level what threads exist in the system and how a logical block is translated all the way down to the correct flash page and then submitted to the NVMe device? It would be greatly appreciated.

Patch-set: 2
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>