nvme: call probe_cb when probing devices in secondary process
If primary process is run with two NVMes and the secondary
is run with just one specified, then the non-specified one
will print an error on attach: `Failed to assign name to NVMe
device`. It will also leak a controller reference. All that
because the name for this drive was never specified in the
config file.

Fix this by calling probe_cb before each attach_cb in
a secondary process. This will tell if the device needs to
be attached or not.

Change-Id: I789ab5f66423a019593f6bff50807972ee54b01e
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>