Update patch set 3
Patch Set 3:

I agree. This is only a test patch and final solution will be a bit different. The is a thread on mailing list regarding this topic and last conclusion was that we still want to do this in data path, but not every I/O - just every X number of I/O. That will be changed soon.

Patch-set: 3
Reviewer: Maciej Szwed <1011262@d5d70762-12d0-45a1-890d-524b12d3f735>