Update patch set 1
Patch Set 1:

Given the great flexibility (along with great difficulty and great confusion) of the ARMv8 AArch64 memory model and associated "dmb" and "dsb"
instructions, it is impossible to choose a set of these instructions to use
that are both safe (i.e. do the sufficient amount of synchronization) AND
optimally efficient (i.e. do only the necessary amount of synchronization).

So my recommendation is to be conservative and choose instruction options
that are guaranteed(?) to be safe.  A given user of this package is always
then free to customize these choices for their specific ARM SoC/processor.
Such a user would of course be expected to be very knowledgeable about the
ARMv8 data synchronization architecture along with detailed understanding
of their HW.  But the average user, won't do this, and probably shouldn't.
Instead they should expect that the SPDK choices are SAFE and will work on
their specific platform.

So this translates into the following choices:  
    For all dsb and dmb instructions, use the following <options>
        1. for read barriers,  always use "ld"
        2. for write barriers,  always use "st"
        3. for read/write barriers, always use "sy"

This is equivalent to saying that as regards to the appropriate "shareable
domain", we always assume the worst/most conservative value of "Full".
So, I think that the spdk_smp_*() barriers should be changed from using
"ishld", "ishst" and "ish" to "ld", "st" and "sy" respectively.

Now for the harder question/issue.  When to use a dmb versus a dsb
instruction?
A "dsb" is a stronger and so potentially slower barrier instruction than a
"dmb", because a "dsb" will wait for the (matching) previous memory
accesses to "COMPLETE", whereas a "dmb" will just make sure that the
(matching) previous memory accesses will be done before the later memory
accesses - i.e. will preserve order between those matching memory accesses
before and those after.

Your reply described the use of an lfence.  This sounds just like a dmb instruction.  So based on my limited understanding of the difference
between the SMP versions and the non-SMP versions,  I would probably use
the "dsb" instruction (with "ld", "st" or "sy") for the ARM SMP barriers.  For the non-SMP ARM barriers, I would probably use the "dmb" instruction
for all three cases (along with "ld", "st" or "sy" as appropriate).

Patch-set: 1
Reviewer: Barry Spinney <1012669@d5d70762-12d0-45a1-890d-524b12d3f735>