bdev_nvme: handle lost controller in reset
We have handling for when pcie NVMe drives are hotplugged, but if the
target application crashes in NVMe-oF, it is possible that we will not
get any notification about the drives being removed from the fabric. In
attempting to reset a "hotplugged" NVMe-oF controller, we attempt to
read some information from it which results in admin commands being sent
over the wire. In the case of a crashed target, these requests simply
never get a response and block the target from shutting down cleanly.

Change-Id: I6265da77d2e67454943ff2827b188fa4fb04df9b
Signed-off-by: Seth Howell <seth.howell@intel.com>