427294 respin
This is a respin of the Jim's [earlier
patch](https://review.gerrithub.io/c/spdk/spdk/+/427294/) that ensures
top-down bdev un-registration, with proper error handling in place of
the assert(s).

I've been running it without a single issue for weeks, enjoying very
helpful differentiation between regular shutdown of my virtual bdev
and hot-remove of its base bdev. This submission aims to subject the
patch with error handling to the standard test suite in hope it passes
and may be merged.

This revision incorporates review comments, avoiding the initial patch
logic of skipping device in the REMOVING state under the iterator loop.

Change-Id: Iafde9558430bc5ce56e8608ef50bcb2b5fbfbf71
Signed-off-by: Andrey Kuzmin <akuzmin@jetstreamsoft.com>