Update patch set 26
Patch Set 26:

On master, 19.01-pre is fine.  If someone builds a package from master, we don't want the version to indicate it's 18.10.1 or any other official release.

The idea is that someone builds an actual package from a tagged release.  So part of the release process is that we have to update the version in the spec file, in addition to usual changes in include/spdk/version.h.

Patch-set: 26