t10dif: Generate and verify T10 DIF for extended LBA.
T10 DIF or DIX will be supported in SPDK throughout, e.g., NVMe driver,
NVMe-oF initiator and target, NVMe block device, malloc block device,
SCSI, iSCSI, FIO plugin, and Perf.

Generic and common APIs to generate and verify T10 DIF will be helpful
for them.

Supporting these APIs in the NVMe driver library was considered first.
However the NVMe driver is not aware of struct iovec, and if these APIs
are in the NVMe driver, iSCSI cannot utilize them. Hence this patch adds
APIs in the utility.

This patch adds the following APIs to generate and verify T10 DIF for
extended LBAs:
- When the SGL buffer doesn't contain DIF field,
  - Generate DIF for each data block and interleave DIF with data block
    into contiguous buffer.
  - Verify DIF for each data block and strip DIF from contiguous buffer
    and write data block into original SGL buffer.
- When the SGL buffer contains DIF field,
  - Generate DIF for each data block and overwrite the DIF field.
  - Verify DIF for each data block.

ISA-L is not utilized in this patch yet.

Change-Id: Ie6588d960113761f10efbf2d2a3cae004af37ce8
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>