t10dif: Generate and verify T10 DIF for SGL payload as byte alignment and granularity
T10 DIF or DIX will be supported in SPDK throughout, e.g., NVMe driver,
NVMe-oF initiator and target, NVMe block device, malloc block device,
SCSI, iSCSI, FIO plugin, and Perf.

Generic and common APIs to generate and verify T10 DIF will be helpful
for them.

Supporting these APIs in the NVMe driver library was considered first.
However the NVMe driver is not aware of struct iovec, and if these APIs
are in the NVMe driver, iSCSI cannot utilize them. Hence this patch adds
APIs in the utility.

This patch adds APIs to generate and verify T10 DIF for SGL payload
as byte alignment and granularity. These APIs are optimized in the path
that each iovec is ELBA granlarity. But the performance is not satisified
until the API in ISA-> is used. Subsequent patches may replace the API to
calculate CRC by the one in ISA-L.

Subsequent patch will add APIs in contiguous buffer case and in T10 DIX,
i.e. separated metadata.

Change-Id: Ie6588d960113761f10efbf2d2a3cae004af37ce8
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>