app: remove the broken g_default_opts
We used this variable to print default options
in app usage output, but the variable was never
initialized. In the usage output, the default
config was always (null) and mem_size was "all
hugepage memory".

To keep things simple, we remove the g_default_opts
entirely and hardcode default values in the app
usage text.

Change-Id: I92401356f87ef734c5d13dfb8be069265c3494ef
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>