Update patch set 1
Patch Set 1: Code-Review-1

The original idea here was that some applications may override the default mem_size - and when a user runs that application and asks for the help output, we want to show what the default mem_size would be for *this* application, not the general SPDK default.  Similar for the config file.  If an application sets the default config_file to "foo.conf" before calling spdk_app_parse_args(), we want the help message to indicate that foo.conf will be the default configuration file if one is not specified.

We do the memcpy in spdk_app_parse_args - this is what initializes g_default_opts.

All that being said - certainly we should have printed out "none" instead of trying to print a NULL pointer; and with your recent memory management changes, instead of "all hugepage memory" it should print "0" - at least on Linux - I see you're fixing that up in the next patch

I'd still prefer that we try to account for the app-specific options when printing out the help message.  Certainly parts of it need to be fixed though.

Patch-set: 1
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1