Update patch set 16
Patch Set 16:

NOTE:  there will be a lot of churn in this over the next 3-5 days as I add a bunch of code to test (basically) the direct interface to compressdev. When reduce lib is added much of this stuff is expected to go away so marking WIP as reviewing at this moment is probably not worth anyones time.  WIll make a note when compressdev API has been basically tested.

Currently, the ISAL based PMD is initializing and starting OK

Patch-set: 16
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>