Update patch set 11
Patch Set 11:

So I left the option to configure --with-reduce but didn't put it in the help.  Left it in so that internally it can be configured for UT even w/o all the compression stuff enabled.

Patch-set: 11
Reviewer: Paul Luse <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Work-in-progress: false