Update patch set 1
Patch Set 1: Code-Review-1

It seems that: 

We have restrict the maximal MaxQueueDepth in the configuration file. And when we using iSCSI initiator to test (we may test bdevperf with iSCSI initiator) we use high queue depth. So I think that here may not be right fix. Can we have some limitation on iSCSI initiator based bdev, we may need to check the supported queue depth and deny the test if we execeed the queue depth from the iSCSI target.

Patch-set: 1
Reviewer: Ziye Yang <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1