[Test]env_dpdk: set hugepage_single_segments to true as default value.
With this default setting, it can save the time
to start the application when the app is specifying
the huge page memory size.

Though we provide the -g parameter, but if this configuration
is set to true as default, which could be better.

This is only for test, this support have some limitations. For
example, it cannot pass the multiple process test.

Change-Id: I2d7e075cc527921398d55b1471c0e188a7032f0b
Signed-off-by: Lizhou Gao <lizhou.gao@ucloud.cn>
Signed-off-by: Ziye Yang <ziye.yang@intel.com>