Update patch set 5
Patch Set 5:

> (2 comments)
 > 
 > I just merged https://review.gerrithub.io/#/c/spdk/spdk/+/432356/.
 > Can you rebase this on top of that and update the shutdown test so
 > that after the initiator is killed, a new initiator is started and
 > runs to completion? That should expose this bug.

Perf still doesn't run to completion in the shutdown test even when based on this patch. We would need to add a mechanism in bdevperf to notice that it is not getting any I/O from the device, and try to reset the drive. Since we don't have that mechanism, bdevperf just keeps polling.
I can look into adding this mechanism to bdevperf.

Patch-set: 5