Update patch set 26
Patch Set 26:

> On master, 19.01-pre is fine. 

Fine for me. So you will merge this for 18.10.1?

About 19.01-pre - this tag is used to craft "Source: " archive link. If I change this to 19.01-pre this will make this link invalid so rpmbuil might not be abble to download source archive and build RPM.

We can ignore this issue in master and just change "Source :" to point to
> https://github.com/spdk/spdk/archive/master.zip

then we can make "Version: 19.01-pre"

Patch-set: 26
Reviewer: Pawel Wodkowski <1011230@d5d70762-12d0-45a1-890d-524b12d3f735>