Update patch set 27
Patch Set 27:

> Patch Set 26:
> 
> > On master, 19.01-pre is fine. 
> 
> Fine for me. So you will merge this for 18.10.1?
> 
> About 19.01-pre - this tag is used to craft "Source: " archive link. If I change this to 19.01-pre this will make this link invalid so rpmbuil might not be abble to download source archive and build RPM.
> 
> We can ignore this issue in master and just change "Source :" to point to
> > https://github.com/spdk/spdk/archive/master.zip
> 
> then we can make "Version: 19.01-pre"


We have additional scripts as part of our build process to ensure the compressed tarball is created dynamically (and named properly based on the version) from an internally maintained copy of the source repo, which is then copied into the _topdir SOURCES directory prior to the invocation of rpmbuild.

Pawel, my understanding is that downloading the source archive from the actual link can only be accomplished via undefining macro '_disable_source_fetch' which should be used with care; i.e. add checksum step to verify.   Some discussion on this topic:   https://stackoverflow.com/questions/33177450/how-do-i-get-rpmbuild-to-download-all-of-the-sources-for-a-particular-spec

Patch-set: 27
Reviewer: Lance Hartmann <1011995@d5d70762-12d0-45a1-890d-524b12d3f735>