autotest_common: fix waitforlisten for FreeBSD
waitforlisten can't be used on FreeBSD as there is no ip command and
netstat output is missing 'Status' column for Unix sockets. So for
non-Linux OS just try to call the get_rpc_methods RPC command.

Second thing: in case the process we are waiting for fail to start
listening on RPC socket the test script will stuck till the build times
out. Fix also this by adding 20s timeout and 0.5s sleep in the while
loop.

Change-Id: Ia8b06af7875b65a7fd8be65cf55e92881f6f95db
Signed-off-by: Pawel Wodkowski <pawelx.wodkowski@intel.com>