test/iscsi_tgt: Port config file from .INI to JSON in the FIO test
Port the config file used in the FIO test from .INI to JSON.

Besides, set_bdev_nvme_hotplug RPC cannot be used easily because
the iSCSI target is not the primary process in the FIO test.

The FIO test doesn't use the hotplug feature and hence remove
enabling hotplug simply.

Change-Id: Idf8809aa829b67a7028f6fa4778a7aae1b431bb8
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>