Update patch set 11
Patch Set 11:

In that later patch where you use this, I'll probably end up suggesting that you put a state variable on the client object that is CONNECTING or ACTIVE and do an if statement, rather than using a function pointer. I think that's both easier to read and better for the branch predictor. So for this patch can you remove the function pointer and we'll discuss adding it back in on the patch that adds the poll function for connecting?

Patch-set: 11
Reviewer: Ben Walker <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>