Update patch set 28
Patch Set 28:

I've asked Lance over IRC to check out:

https://pagure.io/packaging-committee/issue/233

We could use this to save the tar file to the local system with an spdk- prefixed name.  I still think we should get this version committed as-is for master.  We can adopt this scheme when cherry-picking to 18.10 branch, and adjust this on master as well later.

Patch-set: 28
Reviewer: Jim Harris <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>