Update patch set 3
Patch Set 3:

So I was thinking about why this still failed. I will have to add a couple of extra prints tomorrow to confirm it locally, but I am about 95 percent sure that the issue is with non-inline requests (contiguous or sgl) when we have a target that supports less than NVMF_SGL/2 SGL elements (msdbd). The problem is especially bad when the target only supports a single NVMF sgl which both the kernel and spdk do right now.
I am pretty certain that in those instances we cannot split into RDMA SGEs because the NVMF spec requires us to send the rkeys for our buffers across in an NVMe SGL segment element and allow the remote NIC to do the copying. So in the non-inline case where the target only supports a single NVMF sge, we again lose all ability to split along MRs. 
I will do a little research tomorrow to confirm that is the case. If it is, I think we have only one option. Before we create an NVMe request from a bdev I/O, we need to be able to query the target with our buffers(s) address and further split the I/O into multiple requests before sending them to the transport. 
I think that will be the only way to be certain that we work with dynamic memory allocation in every case.

Patch-set: 3
Reviewer: Seth Howell <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>