build: don't include spdk.app.mk in test/cpp_headers
This Makefile doesn't depend on anything in spdk.app.mk,
and provides its own rules that are different than any
of the other applications (including unit tests).

The exception here is the simple "install : app" rule.
Just add that rule to the cpp_headers Makefile instead
of including spdk.app.mk just for that rule.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I684dfba9108dfa7f06d7eb4acecd6806e5797312