Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

Hi Shuhei,

For the current implementation, even we does not split large IO considering a fixed amount of bytes per timeslice, in the next timeslice, there is a compensation to submit less amount of overall bytes. The bandwidth goal is not that relevant to I/O size. As the timeslice is in millisecond, so that when we run the test in seconds, it could be ok with like 5% variation.

Hope this give some information and I think we may add one comment for the reason why we pick a range (0.9 ~ 1.05) to check whether the QoS goal reaches.

Thanks!
Gang

Patch-set: 10
Reviewer: GangCao <1011207@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1