nvme/pcie: don't allow constructing a controller from secondary process
With various possibilities to leak the rte_pci_device in the
primary process, we could technically construct the controller
in secondary. The nvme stack is not prepared for this and
will fail to initialize the device, but nevertheless will leak
the newly created device.

This patch adds an extra check to prevent any controller from
being constructed in secondary process. This will get handy
with the upcoming DPDK 18.11 hotplug functionality.

Change-Id: I772f42b541c5db53310362b6595cebf9a30e8491
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>